refactoring receipts
